+++ src/app/map/map.component.spec.ts
@@ -273,6 +273,80 @@ describe('MapComponent', () => {
         expect(component.stations).toEqual(response);
     });
 
+    it('should call getStreamGages and return list of all gages', () => {
+        const response = [];
+        let mapStreamGageSpy = spyOn(component, 'mapStreamGageResults');
+        let streamgageServiceSpy = spyOn(component.streamgageService, 'getStreamGages').and.returnValue(
+            of(response)
+        );
+        component.streamgagesVisible = true;
+        component.map.setZoom(10);
+        component.loadStreamGages();
+        fixture.detectChanges();
+        expect(component.streamGages).toEqual(response);
+        expect(component.bbox).not.toBeUndefined();
+        expect(streamgageServiceSpy).toHaveBeenCalledWith(component.bbox);
+        expect(mapStreamGageSpy).toHaveBeenCalledWith(component.streamGages, component.streamGageIcon);
+    });
+
+    it('loadStreamGages should not call mapStreamGageResuts if layer is not visible or zoom < 9', () => {
+        let spyOnQueryMethod = spyOn(component, 'mapStreamGageResults').and.callThrough();
+        let streamgageServiceSpy = spyOn(component.streamgageService, 'getStreamGages');
+        component.streamgagesVisible = false;
+        component.map.setZoom(8);
+        component.loadStreamGages();
+        fixture.detectChanges();
+        expect(streamgageServiceSpy).not.toHaveBeenCalledWith(component.bbox);
+        expect(spyOnQueryMethod).not.toHaveBeenCalledWith(component.streamGages, component.streamGageIcon);
+        expect(component.bbox).toBeUndefined();
+    });
+
+    it('should call getSingleGage and return stream gage', () => {
+        const response = [];
+        const siteCode = "07381355";
+        const timeQueryRange = "&startDT=2020-10-06&endDT=2020-10-13";
+        component.streamgageService.getSingleGage(siteCode, timeQueryRange);
+        spyOn(component.streamgageService, 'getSingleGage').and.returnValue(
+            of(response)
+        );
+        spyOn(component, 'queryStreamGageGraph').and.callThrough();
+        fixture.detectChanges();
+        expect(component.singleGage).toEqual(response);
+    });
+
+    it('should call queryStreamGageGraph on marker click', () => {
+        let spyOnQueryMethod = spyOn(component, 'queryStreamGageGraph');
+        let marker = Array.from(document.getElementsByClassName(component.streamgageService.streamGageMarkers));
+        marker.forEach((element) => {
+            element.dispatchEvent(new Event('click'));
+            fixture.detectChanges();
+            expect(spyOnQueryMethod).toHaveBeenCalled();
+        })
+    })
+
+    it('stream gage button should be enabled when zoomed to 9 or higher', () => {
+        component.map.setZoom(10);
+        fixture.detectChanges();
+        expect(document.querySelectorAll<HTMLInputElement>('.leaflet-control input[type="checkbox"]')[4].disabled)
+            .toBeFalse;
+    });
+
+    it('layers should be cleared when zoomed to 8 or lower', () => {
+        component.map.setZoom(8);
+        fixture.detectChanges();
+        expect(component.streamgageService.streamGageMarkers.getLayers()).toEqual([]);
+    });
+
+    it('stream gage button should be disabled but checked and layers cleared when zoomed to lower than 9', () => {
+        component.map.setZoom(8);
+        component.streamgagesVisible = true;
+        fixture.detectChanges();
+        expect(document.querySelectorAll<HTMLInputElement>('.leaflet-control input[type="checkbox"]')[4].disabled)
+            .toBeTrue;
+        expect(component.streamgagesVisible).toBeTrue;
+        expect(component.streamgageService.streamGageMarkers.getLayers()).toEqual([]);
+    });
+
     it('clustering should be disabled in all sites layer when zoomed to 12 or higher', () => {
         component.map.setZoom(12);
         fixture.detectChanges();
@@ -280,13 +354,42 @@ describe('MapComponent', () => {
             .toBeTrue;
     });
 
-    it('AHPS Gage, current warnings, and watches/warnings layers should be removed when map zooms out to 8', () => {
+    it("disableStreamGage should be disable checkbox", () => {
+        component.disableStreamGage();
+        fixture.detectChanges();
+        expect(document.querySelectorAll<HTMLInputElement>('.leaflet-control input[type="checkbox"]')[4].disabled)
+            .toBeTrue;
+    });
+
+    it('queryStreamGageGraph should be called', () => {
+        const e = new L.marker([43.44, -87.75]);
+        const response = [];
+        component.submittedEvent = {
+            event_id: 305, event_name: "2020 Hurricane Delta", event_start_date: "2020-10-06T05:00:00", event_end_date: "2020-10-13T05:00:00"
+        }
+        let singleGageSpy = spyOn(component.streamgageService, 'getSingleGage').and.returnValue(
+            of(response));
+        component.queryStreamGageGraph(e);
+        fixture.detectChanges();
+        expect(singleGageSpy).toHaveBeenCalled();
+        component.submittedEvent = "2020 Hurricane Delta"
+        let graphDiv = document.createElement("div");
+        graphDiv.id = "graphDiv";
+        document.body.appendChild(graphDiv);
+        component.queryStreamGageGraph(e);
+        fixture.detectChanges();
+        expect(singleGageSpy).toHaveBeenCalled();
+    });
+
+    it('AHPS Gage, current warnings, streamgages, and watches/warnings layers should be removed when map zooms out to 8', () => {
         component.ahpsGagesVisible = true;
         component.currWarningsVisible = true;
         component.watchWarnVisible = true;
+        component.streamgagesVisible = true;
         component.map.addLayer(component.AHPSGages);
         component.map.addLayer(component.warnings);
         component.map.addLayer(component.watchesWarnings);
+        component.map.addLayer(component.streamgageService.streamGageMarkers);
         component.map.setZoom(9);
         component.map.setZoom(8);
         component.map.zoom = 8;
@@ -294,6 +397,7 @@ describe('MapComponent', () => {
         expect(component.map.hasLayer(component.AHPSGages)).toBeFalse;
         expect(component.map.hasLayer(component.warnings)).toBeFalse;
         expect(component.map.hasLayer(component.watchesWarnings)).toBeFalse;
+        expect(component.map.hasLayer(component.watchesWarnings)).toBeTrue;
     });
 
     it('there should be as many queries as there are networks', () => {
@@ -410,30 +514,38 @@ describe('MapComponent', () => {
     })
 
     it ('onResize should be called on window resize', () => {
-        let resizeSpy = spyOn(component, 'onResize');
+        component.isClicked = true;
+        let resizeSpy = spyOn(component, 'onResize').and.callThrough();
         window.dispatchEvent(new Event('resize'));
+        fixture.detectChanges();
         expect(resizeSpy).toHaveBeenCalled();
+        if (component.isMobile){
+            expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('none');
+        }else{
+            expect(component.isClicked).toBeFalse;
+            expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('none');
+        }
     });
 
     it ('mobile minimize button should be checked on resize', () => {
+        component.isClicked = false;
         if (component.isMobile){
-            if (component.isClicked){
-                expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('none');
-            }else{
-                expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('flex');
-            }
+            expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('flex');
         }else{
             expect(window.getComputedStyle(document.getElementById('mobile-minimize-button')).display).toEqual('none');
         }
     });
 
-    it ('isClicked should change value', () => {
+    it ('isClicked should change value to true', () => {
+        component.isClicked = false;
         component.openMapFilters();
-        if (component.isClicked){
-            expect(component.isClicked).toBeTruthy();
-        }else{
-            expect(component.isClicked).toBeFalsy();
-        }
+        expect(component.isClicked).toBeTruthy();
+    });
+
+    it ('isClicked should change value to false', () => {
+        component.isClicked = true;
+        component.openMapFilters();
+        expect(component.isClicked).toBeFalsy();
     });
 
     it ('mobile minimize button should be hidden on Map Filters button click', () => {