+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 -   Added map filter button on mobile.  Map filters open to replace the map and filter results.
+-   Added Real-Time Stream Gage layer
 
 ### Changed